feat(BenchmarkScheduler): Адаптировал для работы без DVC
Адаптировал для работы без DVC
